+++ src/Particle/LoadBalance/Par_LB_SendParticleData.cpp
@@ -50,9 +50,9 @@
 //                NRecvPatchTotal, NRecvPatchTotal
 //-------------------------------------------------------------------------------------------------------
 void Par_LB_SendParticleData( const int NParAtt, int *SendBuf_NPatchEachRank, int *SendBuf_NParEachPatch,
-                              long *SendBuf_LBIdxEachPatch, real_par *SendBuf_ParDataEachPatch, const int NSendParTotal,
+                              long *SendBuf_LBIdxEachPatch, real_par *SendBuf_ParDataEachPatch, const long NSendParTotal,
                               int *&RecvBuf_NPatchEachRank, int *&RecvBuf_NParEachPatch, long *&RecvBuf_LBIdxEachPatch,
-                              real_par *&RecvBuf_ParDataEachPatch, int &NRecvPatchTotal, int &NRecvParTotal,
+                              real_par *&RecvBuf_ParDataEachPatch, int &NRecvPatchTotal, long &NRecvParTotal,
                               const bool Exchange_NPatchEachRank, const bool Exchange_LBIdxEachRank,
                               const bool Exchange_ParDataEachRank, Timer_t *Timer, const char *Timer_Comment )
 {
@@ -155,7 +155,7 @@ void Par_LB_SendParticleData( const int NParAtt, int *SendBuf_NPatchEachRank, in
 
 //    send/recv count
       const int *SendPtr = NULL, *RecvPtr = NULL;
-      NRecvParTotal = 0;
+      NRecvParTotal = 0L;
 
       for (int r=0; r<MPI_NRank; r++)
       {
@@ -168,7 +168,7 @@ void Par_LB_SendParticleData( const int NParAtt, int *SendBuf_NPatchEachRank, in
          for (int p=0; p<SendBuf_NPatchEachRank[r]; p++)    SendCount_ParDataEachPatch[r] += SendPtr[p];
          for (int p=0; p<RecvBuf_NPatchEachRank[r]; p++)    RecvCount_ParDataEachPatch[r] += RecvPtr[p];
 
-         NRecvParTotal += RecvCount_ParDataEachPatch[r];
+         NRecvParTotal += (long)RecvCount_ParDataEachPatch[r];
 
          SendCount_ParDataEachPatch[r] *= NParAtt;
          RecvCount_ParDataEachPatch[r] *= NParAtt;
@@ -184,7 +184,7 @@ void Par_LB_SendParticleData( const int NParAtt, int *SendBuf_NPatchEachRank, in
       }
 
 //    reuse the MPI recv buffer declared in LB_GetBufferData for better MPI performance
-      RecvBuf_ParDataEachPatch = (real_par *)LB_GetBufferData_MemAllocate_Recv( NRecvParTotal*NParAtt*sizeof(real_par) );
+      RecvBuf_ParDataEachPatch = (real_par *)LB_GetBufferData_MemAllocate_Recv( NRecvParTotal*(long)NParAtt*sizeof(real_par) );
 
 //    exchange data
       MPI_Alltoallv( SendBuf_ParDataEachPatch, SendCount_ParDataEachPatch, SendDisp_ParDataEachPatch, MPI_GAMER_REAL_PAR,
